+++ test/parser/test_split_quotes.c
@@ -14,7 +14,8 @@
 #include "expander.c"
 #include "expand_variables.c"
 #include "key.c"
-#include "env_var.c"
+#include "get_env_var.c"
+#include "expand_var.c"
 #include "arr_utils.c"
 #include "str_equal.c"
